safari-600.7-branch/LayoutTests/platform/ios-sim/fast/css/hover-update-expected.txt

BUG ID: Bugzilla bug 6821 Fix for 5983 will not always update hover correctly.

To run manually, move your mouse over the black box. You should see a green box. If you see any other color, the test fails.

Only green now
Hover color before test: rgb(255, 255, 0)
FAIL window.getComputedStyle(document.getElementById('b'), null).backgroundColor should be rgb(0, 128, 0). Was rgb(255, 255, 0).
FAIL window.getComputedStyle(document.getElementById('a'), null).backgroundColor should be rgb(0, 128, 0). Was rgb(255, 0, 0).
